+++ tests/optic.spec.ts
@@ -801,6 +801,38 @@ describe('partsOf', () => {
   })
 })
 
+describe('reread', () => {
+  const lens = O.optic<string>().reread(x => x.toUpperCase())
+
+  it('get', () => {
+    const result: string = O.get(lens)('foo')
+    expect(result).toEqual('FOO')
+  })
+
+  it('set - monomorphic', () => {
+    const result: string = O.set(lens)('foo')('')
+    expect(result).toEqual('foo')
+  })
+
+  // reread() is monomorphic
+})
+
+describe('rewrite', () => {
+  const lens = O.optic<string>().rewrite(x => x.toUpperCase())
+
+  it('get', () => {
+    const result: string = O.get(lens)('foo')
+    expect(result).toEqual('foo')
+  })
+
+  it('set - monomorphic', () => {
+    const result: string = O.set(lens)('foo')('')
+    expect(result).toEqual('FOO')
+  })
+
+  // rewrite() is monomorphic
+})
+
 describe('lens/elems', () => {
   type Source = { foo: Array<{ bar: string }>; other: number }
   const source: Source = {